Computer Vision and Home automation for security
Using openCV and android it is able to make out if a person is carrying
a weapon(cosco ball in this case) thereby,shutting down the entire house
by locking mechanism (made using arduino).
All the control passes to android application authenticated only with
the fingerprint of a registered user of that particular
house/institutuion to unlock/lock house.

That application alone has the capablity to act as an RFID for entering
homes. No more stress regarding losing keys or forgetting lock
patttern/pins.